gui_sql/requirements.txt
gradio==6.2.0
huggingface-hub>=0.33.5,<2.0
python-dotenv
tqdm
pandas
gcsfs
google-cloud-storage
numpy
sqlalchemy
#google-cloud-sql
